Create PheWASTable component
